+++ lib/sfdpgen/stress_model.h
@@ -1,3 +1,4 @@
 #pragma once
 
-void stress_model(int dim, SparseMatrix D, double **x, int maxit, int *flag);
+/// @return 0 on success
+int stress_model(int dim, SparseMatrix D, double **x, int maxit);